pyproject.toml
[tool.poetry]
name = "fast-api-auth-template"
version = "0.1.0"
description = "Template for FastAPI with basic user auth"
authors = ["Milos Petrovic <milos.brugi@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fast_api_auth_template"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.82.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
SQLAlchemy = "^1.4.41"
psycopg2 = "^2.9.3"
alembic = "^1.8.1"
python-jose = {extras = ["crypthography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
coverage = "^6.4.4"
requests = "^2.28.1"


[tool.poetry.group.dev.dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
